requirements/base.txt
Django==3.2
djangorestframework==3.12.4
psycopg2>=2.8.6,<2.9.0

# Environment
django-environ==0.4.5

# documentation
pyyaml
uritemplate
drf-yasg